add lock time on locked bitcoin row
