+++ src/inpage/index.ts
@@ -2,9 +2,9 @@ import { StacksProvider } from '@stacks/connect';
 import { BitcoinProvider } from 'sats-connect';
 
 import { XverseProviderInfo } from '@utils/constants';
+import LockedBitcoinMethodsProvider, { LockedBitcoinProvider } from './locked.inpage';
 import SatsMethodsProvider from './sats.inpage';
 import StacksMethodsProvider from './stacks.inpage';
-import StakesMethodsProvider, { StakesProvider } from './stakes.inpage';
 
 declare global {
   interface XverseProviders {
@@ -18,7 +18,8 @@ window.XverseProviders = {
   // @ts-ignore
   StacksProvider: StacksMethodsProvider as StacksProvider,
   BitcoinProvider: SatsMethodsProvider as BitcoinProvider,
-  StakesProvider: StakesMethodsProvider as StakesProvider,
+  // @ts-ignore
+  LockedBitcoinProvider: LockedBitcoinMethodsProvider as LockedBitcoinProvider,
 };
 
 // we inject these in case implementors call the default providers
@@ -56,10 +57,11 @@ try {
 try {
   if (document.currentScript?.dataset.isPriority) {
     Object.defineProperties(window, {
-      StakesProvider: { get: () => StakesMethodsProvider, set: () => {} },
+      LockedBitcoinProvider: { get: () => LockedBitcoinMethodsProvider, set: () => {} },
     });
   } else {
-    window.StakesProvider = StakesMethodsProvider as StakesProvider;
+    // @ts-ignore
+    window.LockedBitcoinProvider = LockedBitcoinMethodsProvider as LockedBitcoinProvider;
   }
 } catch (e) {
   console.log(
